requirements.txt
absl-py==1.4.0
asgiref==3.7.2
attrs==23.1.0
cffi==1.15.1
contourpy==1.1.0
cycler==0.11.0
Django==4.0.3
django-environ==0.11.2
flatbuffers==23.5.26
fonttools==4.42.0
keyboard==0.13.5
kiwisolver==1.4.4
matplotlib==3.7.2
mediapipe==0.10.3
numpy==1.25.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
tzdata==2023.3
sorl-thumbnail==12.10.0
ultralytics==8.0.199
